Add a flag to fixed size text splitter to avoid cutting words at chunk boundaries
